pomocy

+++ backend/endpoints/deleteBorrow.js
@@ -9,14 +9,26 @@ const router = Router();
 router.post("/", async (req, res) => {
 	const id = req.body.id;
 
-	const exist = await Borrows.find({_id: id});
+	const exist = await Borrows.findById(id);
 
+
+	// fuck-proofing the code
 	if(!exist) {
 		error("ale jak?????", res);
 		return;
 	}
 
-	// if(exist.)
+	if(exist.received) {
+		error("za pozno", res, 418);
+		return;
+	}
+
+	if(exist.user !== req.user._id) {
+		error("ej ej ej", res, 401);
+		return;
+	}
+
+	
 
 	const delete_borrow = await Borrows.deleteOne({_id: id});
 
@@ -26,7 +38,7 @@ router.post("/", async (req, res) => {
 	}
 	else {
 		res.status(200).json({confirm: "yes"});
-		return
+		return;
 	}
 	
 });
